added changes to the controller and json data
