+++ controllers/api/index.js
@@ -4,7 +4,6 @@ const workoutRoutes = require('./workoutRoutes');
 const bodypartsRoutes = require ('./bodypartsRoute');
 
 router.use('./users', userRoutes);
-router.use('./bodyparts', bodypartsRoutes);
-router.use('./workout', workoutRoutes);
+
 
 module.exports = router;
